tests: make some metrics helpers reusable
